complete bundle and merge server files
